+++ packages/manager/apps/okms/src/common/components/okms-dashboard/general-informations-tile/items/IdTileItem.component.tsx
@@ -1,6 +1,7 @@
-import React from 'react';
-import { ManagerTile, Clipboard } from '@ovh-ux/manager-react-components';
 import { OKMS } from '@key-management-service/types/okms.type';
+
+import { Clipboard, ManagerTile } from '@ovh-ux/manager-react-components';
+
 import { ID_LABEL } from '@/constants';
 
 type IdTileItemProps = {