+++ packages/manager/apps/okms/src/common/components/okms-dashboard/okms-domain-dashboard-tiles/OkmsDomainDashboardTiles.constants.ts
@@ -1,0 +1,3 @@
+export const OKMS_DOMAIN_DASHBOARD_TILES_TEST_IDS = {
+  okmsDashboardTiles: 'okms-dashboard-tiles',
+};