+++ packages/manager/apps/okms/src/modules/key-management-service/components/guide/guide-quick-start/guideQuickStart.constants.ts
@@ -1,0 +1,20 @@
+import { GuideLinks } from '@/common/utils/guides/guideLinks.type';
+
+export const GUIDES_QUICK_START: GuideLinks = {
+  ASIA: 'https://help.ovhcloud.com/csm/asia-kms-quick-start?id=kb_article_view&sysparm_article=KB0063365',
+  AU: 'https://help.ovhcloud.com/csm/en-au-kms-quick-start?id=kb_article_view&sysparm_article=KB0063364',
+  CA: 'https://help.ovhcloud.com/csm/en-ca-kms-quick-start?id=kb_article_view&sysparm_article=KB0063371',
+  DE: 'https://help.ovhcloud.com/csm/de-kms-quick-start?id=kb_article_view&sysparm_article=KB0063353',
+  ES: '/https://help.ovhcloud.comcsm/es-es-kms-quick-start?id=kb_article_view&sysparm_article=KB0063370',
+  FR: 'https://help.ovhcloud.com/csm/fr-kms-quick-start?id=kb_article_view&sysparm_article=KB0063369',
+  GB: 'https://help.ovhcloud.com/csm/en-gb-kms-quick-start?id=kb_article_view&sysparm_article=KB0063380',
+  IE: 'https://help.ovhcloud.com/csm/en-ie-kms-quick-start?id=kb_article_view&sysparm_article=KB0063362',
+  IT: 'https://help.ovhcloud.com/csm/it-kms-quick-start?id=kb_article_view&sysparm_article=KB0063381',
+  PL: 'https://help.ovhcloud.com/csm/pl-kms-quick-start?id=kb_article_view&sysparm_article=KB0063385',
+  PT: 'https://help.ovhcloud.com/csm/pt-kms-quick-start?id=kb_article_view&sysparm_article=KB0063388',
+  QC: 'https://help.ovhcloud.com/csm/fr-ca-kms-quick-start?id=kb_article_view&sysparm_article=KB0063373',
+  SG: 'https://help.ovhcloud.com/csm/en-sg-kms-quick-start?id=kb_article_view&sysparm_article=KB0063374',
+  US: 'https://support.us.ovhcloud.com/hc/en-us/articles/31481250033811-Getting-Started-with-OVHcloud-Key-Management-Service-KMS',
+  WE: 'https://help.ovhcloud.com/csm/en-kms-quick-start?id=kb_article_view&sysparm_article=KB0063366',
+  WS: 'https://help.ovhcloud.com/csm/es-es-kms-quick-start?id=kb_article_view&sysparm_article=KB0063370',
+};