+++ packages/manager/apps/okms/src/modules/key-management-service/components/service-key/create/serviceKeyOperationCheckbox.tsx
@@ -1,21 +1,19 @@
-import React from 'react';
+import { useServiceKeyOperationsTranslations } from '@key-management-service/hooks/service-key/useServiceKeyOperationsTranslations';
+import { OkmsServiceKeyReferenceOperations } from '@key-management-service/types/okmsServiceKeyReference.type';
 import { useTranslation } from 'react-i18next';
+
 import {
   ODS_TEXT_PRESET,
-  OdsCheckboxCustomEvent,
   OdsCheckboxChangeEventDetail,
+  OdsCheckboxCustomEvent,
   OdsCheckbox as OdsCheckboxType,
 } from '@ovhcloud/ods-components';
 import { OdsCheckbox, OdsText } from '@ovhcloud/ods-components/react';
-import { useServiceKeyOperationsTranslations } from '@key-management-service/hooks/service-key/useServiceKeyOperationsTranslations';
-import { OkmsServiceKeyReferenceOperations } from '@key-management-service/types/okmsServiceKeyReference.type';
 
 type TServiceKeyOperationCheckbox = {
   name: string;
   operation: OkmsServiceKeyReferenceOperations;
-  onOdsChange?: (
-    e: OdsCheckboxCustomEvent<OdsCheckboxChangeEventDetail>,
-  ) => void;
+  onOdsChange?: (e: OdsCheckboxCustomEvent<OdsCheckboxChangeEventDetail>) => void;
 } & Partial<OdsCheckboxType>;
 
 export const ServiceKeyOperationCheckbox = ({
@@ -23,9 +21,7 @@ export const ServiceKeyOperationCheckbox = ({
   ...props
 }: TServiceKeyOperationCheckbox) => {
   const { t } = useTranslation('key-management-service/serviceKeys');
-  const translatedOperations = useServiceKeyOperationsTranslations(
-    operation.value,
-  );
+  const translatedOperations = useServiceKeyOperationsTranslations(operation.value);
   const key = operation.value.join('_');
 
   return (
@@ -36,9 +32,7 @@ export const ServiceKeyOperationCheckbox = ({
           {translatedOperations.join(' / ')}
         </OdsText>
         <OdsText preset={ODS_TEXT_PRESET.caption}>
-          {t(
-            `key_management_service_service-keys_create_crypto_field_usage_description_${key}`,
-          )}
+          {t(`key_management_service_service-keys_create_crypto_field_usage_description_${key}`)}
         </OdsText>
       </label>
     </div>