+++ packages/manager/apps/okms/src/modules/key-management-service/mocks/feature-availability/feature-availability.handler.ts
@@ -1,4 +1,5 @@
 import { Handler } from '@ovh-ux/manager-core-test-utils';
+
 import { featuresAvailabilityMock } from './feature-availability.mock';
 
 export type GetFeatureAvailabilityMocksParams = {