+++ packages/manager/apps/okms/src/modules/secret-manager/mocks/secret-config-okms/secretConfigOkms.handler.ts
@@ -1,6 +1,8 @@
 import { Handler } from '@ovh-ux/manager-core-test-utils';
-import { mockSecretConfigOkms } from './secretConfigOkms.mock';
+
 import { buildMswResponseMock } from '@/common/utils/tests/msw';
+
+import { mockSecretConfigOkms } from './secretConfigOkms.mock';
 
 // GET Secret Config
 export const getSecretConfigErrorMessage = 'get-secret-config-error-message';
@@ -17,9 +19,32 @@ export const getSecretConfigOkmsMock = ({
     response: buildMswResponseMock({
       data: mockSecretConfigOkms,
       errorMessage: getSecretConfigErrorMessage,
-      isError: isSecretConfigKO,
+      isError: isSecretConfigKO ?? false,
     }),
     status: isSecretConfigKO ? 500 : 200,
     api: 'v2',
+  },
+];
+
+// PUT Secret Config
+export const updateSecretConfigErrorMessage = 'put-secret-config-error-message';
+
+export type UpdateSecretConfigOkmsMockParams = {
+  isUpdateSecretConfigKO?: boolean;
+};
+
+export const updateSecretConfigOkmsMock = ({
+  isUpdateSecretConfigKO,
+}: UpdateSecretConfigOkmsMockParams): Handler[] => [
+  {
+    url: '/okms/resource/:okmsId/secretConfig',
+    response: buildMswResponseMock({
+      data: mockSecretConfigOkms,
+      errorMessage: updateSecretConfigErrorMessage,
+      isError: isUpdateSecretConfigKO,
+    }),
+    status: isUpdateSecretConfigKO ? 500 : 200,
+    api: 'v2',
+    method: 'put',
   },
 ];