+++ packages/manager/apps/okms/src/modules/secret-manager/pages/drawers/create-version-drawer/CreateVersionDrawer.page.tsx
@@ -1,20 +1,26 @@
-import React, { Suspense } from 'react';
-import { Drawer } from '@ovh-ux/manager-react-components';
-import { useTranslation } from 'react-i18next';
-import { useNavigate, useParams } from 'react-router-dom';
-import { OdsMessage } from '@ovhcloud/ods-components/react';
+import { Suspense } from 'react';
+
+import { useNavigate } from 'react-router-dom';
+
 import { useSecretWithData } from '@secret-manager/data/hooks/useSecret';
 import { useSecretSmartConfig } from '@secret-manager/hooks/useSecretSmartConfig';
 import { decodeSecretPath } from '@secret-manager/utils/secretPath';
-import { LocationPathParams } from '@secret-manager/routes/routes.constants';
-import { CreateVersionDrawerForm } from './CreateVersionDrawerForm.component';
+import { useTranslation } from 'react-i18next';
+
+import { OdsMessage } from '@ovhcloud/ods-components/react';
+
+import { Drawer } from '@ovh-ux/manager-react-components';
+
+import { useRequiredParams } from '@/common/hooks/useRequiredParams';
+
 import { CREATE_VERSION_DRAWER_TEST_IDS } from './CreateVersionDrawer.constants';
+import { CreateVersionDrawerForm } from './CreateVersionDrawerForm.component';
 
 export default function CreateVersionDrawerPage() {
   const navigate = useNavigate();
   const { t } = useTranslation('secret-manager');
 
-  const { okmsId, secretPath } = useParams<LocationPathParams>();
+  const { okmsId, secretPath } = useRequiredParams('okmsId', 'secretPath');
 
   const {
     data: secret,
@@ -49,7 +55,7 @@ export default function CreateVersionDrawerPage() {
             {error?.response?.data?.message}
           </OdsMessage>
         )}
-        {!error && secret && (
+        {!error && secret && secretConfig && (
           <CreateVersionDrawerForm
             secret={secret}
             okmsId={okmsId}